Cleaned up AnalyzeVotes folder
